Add files via upload
This program collects two numbers from the user, and show both the sum, difference and comparison of the two numbers.  Instead of the generic user entry request, I created this sub-program to strip out non-numeric entries. The non-numeric characters are simply stripped off, leaving only integers. If there is no integer entry, the number 0 is used.